id change

+++ src/main/java/Classes/Contact.java
@@ -1,14 +1,14 @@
 package Classes;
 
 public class Contact extends Person{
-    private static long id;
+    private static int id;
 
     public Contact(String name, long phoneNumber, String email) {
         super(name, phoneNumber, email);
         id++;
     }
 
-    public static long getId() {
+    public static int getId() {
         return id;
     }
 }
